glossary/requirements.txt
grpcio==1.60.0
protobuf==4.25.0
pytest==7.4.0